Analysis: Replaced Word documents with GitHub Markdown
Using Word documents in Git is a bad idea anyway, as it is impossible (?)
to view the diffs of the files on GitHub. Also it creates these shadow files
that are marked with a tilde, which are annoying to remove every time.

Using GitHub Markdown for all documentation from now on.
(Besides, can use gitprint.com to convert to printable PDF)