Graph: Add a semi-functional graph
A basic graph now loads on the Graph tab. I have added a data source class after realising the MainWindowController shouldn't really store all the data source methods for the graph.
Axes are not labelled, they won't be; the idea is to display the weight/date at each point individually.

User can scroll and pinch to zoom to interact with the graph.